Remove deprecated platforms long time ago (Legacy Slack and Legacy Teams)
