--- helm/botkube/templates/ingress.yaml
@@ -1,42 +1,0 @@
-{{ if .Values.ingress.create }}
-kind: Ingress
-apiVersion: networking.k8s.io/v1
-metadata:
-  name: {{ include "botkube.fullname" . }}
-  labels:
-    app.kubernetes.io/name: {{ include "botkube.name" . }}
-    helm.sh/chart: {{ include "botkube.chart" . }}
-    app.kubernetes.io/instance: {{ .Release.Name }}
-    app.kubernetes.io/managed-by: {{ .Release.Service }}
-    app: botkube
-  annotations:
-    {{- if .Values.ingress.annotations }}
-{{ toYaml .Values.ingress.annotations | indent 4 }}
-    {{- end }}
-spec:
-{{- if .Values.ingress.tls.enabled }}
-  tls:
-  - hosts:
-    - {{ .Values.ingress.host }}
-    secretName: {{ .Values.ingress.tls.secretName }}
-{{- end }}
-  rules:
-  - http:
-      paths:
-
-        {{- range .Values.communications }}
-        {{- if .teams.enabled }}
-        - path: {{ .teams.messagePath }}
-          pathType: Prefix
-          backend:
-            service:
-              name: {{ include "botkube.fullname" $ }}
-              port:
-                number: {{ .teams.port }}
-        {{- end }}
-        {{- end }}
-
-  {{- if .Values.ingress.host }}
-    host: {{ .Values.ingress.host }}
-    {{- end }}
-{{- end -}}
