+++ helm/botkube/templates/service.yaml
@@ -1,4 +1,4 @@
-{{- if or .Values.serviceMonitor.enabled (include "botkube.communication.team.enabled" $) (.Values.settings.lifecycleServer.enabled ) (.Values.plugins.incomingWebhook.enabled) }}
+{{- if or .Values.serviceMonitor.enabled (.Values.settings.lifecycleServer.enabled ) (.Values.plugins.incomingWebhook.enabled) }}
 apiVersion: v1
 kind: Service
 metadata:
@@ -26,12 +26,6 @@ spec:
   - name: {{ .Values.service.name }}
     port: {{ .Values.service.port }}
     targetPort: {{ .Values.service.targetPort }}
-  {{- end }}
-  {{- range $key, $val := .Values.communications }}
-  {{- if .teams.enabled }}
-  - name: {{ $key | quote }}
-    port: {{ $val.teams.port }}
-  {{- end }}
   {{- end }}
   selector:
     app: botkube
