+++ internal/analytics/testdata/TestSegmentReporter_Run.json
@@ -18,7 +18,7 @@
 							},
 							"error": null,
 							"integrationType": "bot",
-							"platform": "slack",
+							"platform": "cloudSlack",
 							"pluginName": "botkube/kubernetes",
 							"success": true
 						},
@@ -77,7 +77,7 @@
 							},
 							"error": null,
 							"integrationType": "bot",
-							"platform": "slack",
+							"platform": "cloudSlack",
 							"pluginName": "botkube/kubernetes",
 							"success": true
 						}
