+++ pkg/bot/teams_cloud_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/infracloudio/msbotbuilder-go/schema"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"github.com/kubeshop/botkube/pkg/execute/command"
 )
@@ -51,6 +52,47 @@ func TestExtractExplicitOrigin(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			gotOrigin := cloudTeam.mapToCommandOrigin(tc.givenAct)
 			assert.Equal(t, tc.wantOrigin, gotOrigin)
+		})
+	}
+}
+
+func TestTeams_TrimBotMention(t *testing.T) {
+	/// given
+	botName := "Botkube"
+	testCases := []struct {
+		Name               string
+		Input              string
+		ExpectedTrimmedMsg string
+	}{
+		{
+			Name:               "Mention",
+			Input:              "<at>Botkube</at> get pods",
+			ExpectedTrimmedMsg: " get pods",
+		},
+		{
+			Name:               "Not at the beginning",
+			Input:              "Not at the beginning <at>Botkube</at> get pods",
+			ExpectedTrimmedMsg: "Not at the beginning <at>Botkube</at> get pods",
+		},
+		{
+			Name:               "Different mention",
+			Input:              "<at>bootkube</at> get pods",
+			ExpectedTrimmedMsg: "<at>bootkube</at> get pods",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			botMentionRegex, err := teamsBotMentionRegex(botName)
+			require.NoError(t, err)
+			b := &CloudTeams{botMentionRegex: botMentionRegex}
+			require.NoError(t, err)
+
+			// when
+			actualTrimmedMsg := b.trimBotMention(tc.Input)
+
+			// then
+			assert.Equal(t, tc.ExpectedTrimmedMsg, actualTrimmedMsg)
 		})
 	}
 }
