+++ test/e2e/migration_test.go
@@ -36,8 +36,6 @@ const (
 	--set communications.default-group.discord.token=%s \
 	--set settings.clusterName=%s \
 	--set executors.k8s-default-tools.botkube/kubectl.enabled=true \
-	--set executors.k8s-default-tools.botkube/helm.enabled=true \
-	--set executors.bins-management.botkube/exec.config.templates[0].ref=github.com/kubeshop/botkube//cmd/executor/exec/templates?ref=main \
 	--set analytics.disable=true \
 	--set image.tag=v9.99.9-dev \
 	--set plugins.repositories.botkube.url=https://storage.googleapis.com/botkube-plugins-latest/plugins-index.yaml \
@@ -270,23 +268,7 @@ func assertAliases(t *testing.T, actual []*gqlModel.Alias) {
 			Command:     "kubectl",
 			Deployments: nil,
 		},
-		{
-			ID:          "",
-			Name:        "x",
-			DisplayName: "Exec alias",
-			Command:     "exec",
-			Deployments: nil,
-		},
-		{
-			ID:          "",
-			Name:        "chatgpt",
-			DisplayName: "Doctor alias",
-			Command:     "doctor",
-			Deployments: nil,
-		},
 	}
-
-	assert.Len(t, actual, 4)
 
 	// trim ID and deployments
 	for i := range actual {
@@ -313,14 +295,9 @@ func assertPlatforms(t *testing.T, actual *gqlModel.Platforms, appCfg MigrationC
 						Sources: []string{
 							"k8s-err-events",
 							"k8s-recommendation-events",
-							"k8s-err-events-with-ai-support",
-							"argocd",
 						},
 						Executors: []string{
 							"k8s-default-tools",
-							"bins-management",
-							"ai",
-							"flux",
 						},
 					},
 					NotificationsDisabled: ptr.FromType(false),
@@ -330,7 +307,6 @@ func assertPlatforms(t *testing.T, actual *gqlModel.Platforms, appCfg MigrationC
 	}
 
 	assert.NotNil(t, actual, 1)
-	assert.Len(t, actual.Discords, 1)
 
 	// trim ignored fields
 	for i := range actual.Discords {
@@ -399,114 +375,6 @@ func assertPlugins(t *testing.T, actual []*gqlModel.Plugin) {
 			Configuration:     "{\"defaultNamespace\":\"default\"}",
 			Enabled:           true,
 			Rbac:              defaultRBAC,
-		},
-		{
-			Name:              "botkube/helm",
-			DisplayName:       "botkube/helm",
-			Type:              "EXECUTOR",
-			ConfigurationName: "k8s-default-tools",
-			Configuration:     "{\"defaultNamespace\":\"default\",\"helmCacheDir\":\"/tmp/helm/.cache\",\"helmConfigDir\":\"/tmp/helm/\",\"helmDriver\":\"secret\"}",
-			Enabled:           true,
-			Rbac:              defaultRBAC,
-		},
-		{
-			Name:              "botkube/argocd",
-			DisplayName:       "botkube/argocd",
-			Type:              "SOURCE",
-			ConfigurationName: "argocd",
-			Configuration:     "{\"argoCD\":{\"notificationsConfigMap\":{\"name\":\"argocd-notifications-cm\",\"namespace\":\"argocd\"},\"uiBaseUrl\":\"http://localhost:8080\"},\"defaultSubscriptions\":{\"applications\":[{\"name\":\"guestbook\",\"namespace\":\"argocd\"}]}}",
-			Enabled:           false,
-			Rbac: &gqlModel.Rbac{
-				User: defaultRBAC.User,
-				Group: &gqlModel.GroupPolicySubject{
-					Type: defaultRBAC.Group.Type,
-					Static: &gqlModel.GroupStaticSubject{
-						Values: []string{"argocd"},
-					},
-					Prefix: defaultRBAC.Group.Prefix,
-				},
-			},
-		},
-		{
-			Name:              "botkube/keptn",
-			DisplayName:       "botkube/keptn",
-			Type:              "SOURCE",
-			ConfigurationName: "keptn",
-			Configuration:     "{\"log\":{\"level\":\"info\"},\"project\":\"\",\"service\":\"\",\"token\":\"\",\"url\":\"http://api-gateway-nginx.keptn.svc.cluster.local/api\"}",
-			Enabled:           false,
-			Rbac: &gqlModel.Rbac{
-				User: defaultRBAC.User,
-				Group: &gqlModel.GroupPolicySubject{
-					Type: gqlModel.PolicySubjectTypeEmpty,
-					Static: &gqlModel.GroupStaticSubject{
-						Values: []string{},
-					},
-					Prefix: nil,
-				},
-			},
-		},
-		{
-			Name:              "botkube/exec",
-			DisplayName:       "botkube/exec",
-			Type:              "EXECUTOR",
-			ConfigurationName: "bins-management",
-			Configuration:     "{\"templates\":[{\"ref\":\"github.com/kubeshop/botkube//cmd/executor/exec/templates?ref=main\"}]}",
-			Enabled:           false,
-			Rbac:              defaultRBAC,
-		},
-		{
-			Name:              "botkube/doctor",
-			DisplayName:       "botkube/doctor",
-			Type:              "EXECUTOR",
-			ConfigurationName: "ai",
-			Configuration:     "{\"apiBaseUrl\":\"\",\"apiKey\":\"\",\"defaultEngine\":\"\",\"organizationID\":\"\",\"userAgent\":\"\"}",
-			Enabled:           false,
-			Rbac:              defaultRBAC,
-		},
-		{
-			Name:              "botkube/flux",
-			DisplayName:       "botkube/flux",
-			Type:              "EXECUTOR",
-			ConfigurationName: "flux",
-			Configuration:     "{\"github\":{\"auth\":{\"accessToken\":\"\"}},\"log\":{\"level\":\"info\"}}",
-			Enabled:           false,
-			Rbac: &gqlModel.Rbac{
-				User: defaultRBAC.User,
-				Group: &gqlModel.GroupPolicySubject{
-					Type: gqlModel.PolicySubjectTypeStatic,
-					Static: &gqlModel.GroupStaticSubject{
-						Values: []string{"botkube-plugins-default", "flux-read-patch"},
-					},
-					Prefix: defaultRBAC.Group.Prefix,
-				},
-			},
-		},
-		{
-			Name:              "botkube/kubernetes",
-			DisplayName:       "Kubernetes Errors with AI support",
-			Type:              "SOURCE",
-			ConfigurationName: "k8s-err-events-with-ai-support",
-			Configuration:     "{\"event\":{\"types\":[\"error\"]},\"extraButtons\":[{\"button\":{\"commandTpl\":\"doctor --resource={{ .Kind | lower }}/{{ .Name }} --namespace={{ .Namespace }} --error={{ .Reason }} --bk-cmd-header='AI assistance'\",\"displayName\":\"Get Help\"},\"enabled\":true,\"trigger\":{\"type\":[\"error\"]}}],\"namespaces\":{\"include\":[\".*\"]},\"resources\":[{\"type\":\"v1/pods\"},{\"type\":\"v1/services\"},{\"type\":\"networking.k8s.io/v1/ingresses\"},{\"event\":{\"message\":{\"exclude\":[\".*nf_conntrack_buckets.*\"]}},\"type\":\"v1/nodes\"},{\"type\":\"v1/namespaces\"},{\"type\":\"v1/persistentvolumes\"},{\"type\":\"v1/persistentvolumeclaims\"},{\"type\":\"v1/configmaps\"},{\"type\":\"rbac.authorization.k8s.io/v1/roles\"},{\"type\":\"rbac.authorization.k8s.io/v1/rolebindings\"},{\"type\":\"rbac.authorization.k8s.io/v1/clusterrolebindings\"},{\"type\":\"rbac.authorization.k8s.io/v1/clusterroles\"},{\"type\":\"apps/v1/deployments\"},{\"type\":\"apps/v1/statefulsets\"},{\"type\":\"apps/v1/daemonsets\"},{\"type\":\"batch/v1/jobs\"}]}",
-			Enabled:           false,
-			Rbac:              defaultRBAC,
-		},
-		{
-			Name:              "botkube/prometheus",
-			DisplayName:       "botkube/prometheus",
-			Type:              "SOURCE",
-			ConfigurationName: "prometheus",
-			Configuration:     "{\"alertStates\":[\"firing\",\"pending\",\"inactive\"],\"ignoreOldAlerts\":true,\"log\":{\"level\":\"info\"},\"url\":\"http://localhost:9090\"}",
-			Enabled:           false,
-			Rbac: &gqlModel.Rbac{
-				User: defaultRBAC.User,
-				Group: &gqlModel.GroupPolicySubject{
-					Type: gqlModel.PolicySubjectTypeEmpty,
-					Static: &gqlModel.GroupStaticSubject{
-						Values: []string{},
-					},
-					Prefix: nil,
-				},
-			},
 		},
 	}
